Updated Tables/Views

+++ IspDataDb/dbo/Tables/SecurityGroup.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[SecurityGroup] (
-    [IdNo]                 SMALLINT      IDENTITY (1, 1) NOT NULL,
+    [IdNo]                 INT           IDENTITY (1, 1) NOT NULL,
     [SecurityGroupName]    VARCHAR (50)  NULL,
     [ParentIdNo]           SMALLINT      NULL,
     [Notes]                VARCHAR (100) NULL,
@@ -9,6 +9,8 @@
     CONSTRAINT [PK_IDNo] PRIMARY KEY CLUSTERED ([IdNo] ASC),
     CONSTRAINT [IX_SecurityGroupName] UNIQUE NONCLUSTERED ([SecurityGroupName] ASC)
 );
+
+
